requirements.txt
langchain==0.3.27
langchain_community==0.3.27
langchain_groq==0.3.6
crewai
crewai[tools] 
duckduckgo-search==8.1.1 
python-dotenv==1.1.1 
streamlit==1.47.1